+++ js/data.js
@@ -1,0 +1,95 @@
+
+import {getRandomPositiveFloat, getRandomPositiveInteger, getRandomArrayElement, getRandomArray} from './util.js';
+
+const MIN_PRICE = 1000;
+const MAX_PRICE = 1000000;
+
+const MIN_ROOMS = 1;
+const MAX_ROOMS = 5;
+
+const MIN_GUEST = 1;
+const MAX_GUEST = 20;
+
+const MIN_AXIS_X = 35.65000;
+const MAX_AXIS_X = 35.70000;
+
+const MIN_AXIS_Y = 139.70000;
+const MAX_AXIS_Y = 139.80000;
+const AXIS_FLOAT = 5;
+
+const ADS_COUNT = 10;
+
+const TITLES = [
+  'Уютная квартира в центре Рима',
+  'Роскошный дом у моря',
+  'Атмосферный лофт',
+  'Комната в коммунальной квартире',
+];
+
+const TYPES = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+  'hotel',
+];
+
+const CHECK_TIME = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+
+const FEATURES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+];
+
+const DESCRIPTIONS = [
+  'Снять комнату в Риме с ванной комнатой в центре посуточно, двухместная кровать',
+  'Дом не далеко от пляжа',
+  'Квартира в центре города, без воды',
+];
+
+const PHOTOS = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
+];
+
+const generateAds = (index) => {
+  const locationLat = getRandomPositiveFloat (MIN_AXIS_X, MAX_AXIS_X, AXIS_FLOAT);
+  const locationLng= getRandomPositiveFloat (MIN_AXIS_Y, MAX_AXIS_Y, AXIS_FLOAT);
+
+  return {
+    author: {
+      avatar: `img/avatars/user${index < 10 ? `0${index}` : index}.png`,
+    },
+    offer : {
+      title: getRandomArrayElement(TITLES),
+      address: {
+        lat: locationLat,
+        lng: locationLng,
+      },
+      price: getRandomPositiveInteger(MIN_PRICE, MAX_PRICE),
+      type: getRandomArrayElement(TYPES),
+      rooms: getRandomPositiveInteger(MIN_ROOMS, MAX_ROOMS),
+      guests: getRandomPositiveInteger(MIN_GUEST, MAX_GUEST),
+      checkin: getRandomArrayElement(CHECK_TIME),
+      checkout: getRandomArrayElement(CHECK_TIME),
+      features: getRandomArray(FEATURES),
+      photos: getRandomArray(PHOTOS),
+      description: getRandomArrayElement(DESCRIPTIONS),
+    },
+    location : {
+      lat: locationLat,
+      lng: locationLng,
+    },
+  };
+};
+
+export {generateAds, ADS_COUNT};
